requirements.txt
pandas==2.2.3
chromadb==1.3.5
groq==0.33.0
python-dotenv==1.1.0
semantic-router==0.1.12
